+++ packages/expo-audio/src/AudioModule.web.ts
@@ -218,6 +218,10 @@ export class AudioPlayerWeb
     getStatusFromMedia(this.media, this.id);
   }
 
+  setActiveForLockScreen(active: boolean, metadata: Record<string, any>): void {}
+  updateLockScreenMetadata(metadata: Record<string, any>): void {}
+  clearLockScreenControls(): void {}
+
   _createMediaElement(): HTMLAudioElement {
     const newSource = getSourceUri(this.src);
     const media = new Audio(newSource);